Release 1.8.0
For 1.8.0, a new feature is released for both Group call and Direct call features respectively.
For the Group call feature, you can now add and manage custom items to store additional information for a room.
Here are the details of the update:
    * Added `customItems` in `Room`.
    * Added `customItems` in `RoomParams`.
    * Added `updateCustomItems(customItems:completionHandler)` and `deleteCustomItems(customItemKeys:completionHandler:)` in `Room`.
    * Added `didUpdateCustomItems(updatedKeys:)` and `didDeleteCustomItems(deletedKeys:)` in `RoomDelegate`.
For the Direct call feature, you can now hold and resume calls which allows you to accept an incoming call or switch between calls.
Here are the details of the update:
    * Added `hold(completionHandler:)` and `unhold(force:completionHandler)` in `DirectCall`.
    * Added `isOnHold` in `DirectCall`.
    * Added `holdActiveCall` in `DialParams` and `AcceptParams`.
    * Added `didUserHoldStatusChange(call:isLocalUser:isUserOnHold:)` in `DirectCallDelegate`.
* Added `getOngoingCalls()` in `SendBirdCall` to retrieve a list of ongoing Direct Calls in the Calls SDK.
* API reference is updated for all platforms.
* Fixed an issue where some characters in `userId` were incorrectly encoded during authentication.
* Added support for Swift Package Manager.
* Improved stability.